// angular/src/app/models/prere.ts
export interface Prere {    
    email?: string;
    id: string; 
    apellido?: string;
    nombre: string;
    numCuil?: number;
    numDoc?: number;
    nacionalidad?: string;
    fecha_nacimiento?: number;
    domicilio?: string;
    cuidad?: string;
    provincia?: string;
    telefono?: number;
    carrera:string,
    anioInscripcion?:number;
    fotoDNI: string;
    nivelUser: string;
}